feat: adicionando codigo e mesa

+++ src/main/java/com/abdiel/restauranteRest/RestauranteApp/dtos/response/PedidoDTO.java
@@ -3,7 +3,7 @@ package com.abdiel.restauranteRest.RestauranteApp.dtos.response;
 import java.time.LocalDateTime;
 
 public record PedidoDTO(
-        int codigo,
+        Long codigo,
         int mesa,
         String nome,
         LocalDateTime data
